fix ui flickering
