--- tests/e2e/auth.spec.ts
@@ -1,254 +1,0 @@
-import { expect, test } from '@playwright/test';
-
-const BASE_URL = process.env.BASE_URL ?? 'http://localhost:8787';
-const TEST_USERNAME = process.env.TEST_USERNAME ?? '';
-const TEST_PASSWORD = process.env.TEST_PASSWORD ?? '';
-const AUTH_EMAIL_SELECTOR = process.env.PLAYWRIGHT_AUTH_EMAIL_SELECTOR ?? 'input[name="email"]';
-const AUTH_PASSWORD_SELECTOR = process.env.PLAYWRIGHT_AUTH_PASSWORD_SELECTOR ?? 'input[name="password"]';
-const AUTH_SUBMIT_SELECTOR = process.env.PLAYWRIGHT_AUTH_SUBMIT_SELECTOR ?? 'button[name="intent"]:not([data-method])';
-const AUTH_CONTINUE_SELECTOR = process.env.PLAYWRIGHT_AUTH_CONTINUE_SELECTOR ?? 'button:has-text("Continue")';
-
-function isAuthKitUrl(url: URL): boolean {
-	return url.hostname.includes('authkit.app') || url.pathname.includes('/auth/signin');
-}
-
-	test.describe('Authentication Flow', () => {
-		test.beforeEach(async ({ context }) => {
-			const storageState = await context.storageState();
-			if (storageState.cookies.length === 0) {
-				await context.clearCookies();
-			}
-		});
-
-	test('unauthenticated user - verify initial state on home page', () => {
-		test.skip(true, 'This test requires manual browser testing - the auth state is determined client-side');
-	});
-
-	test('protected route redirects to signin', async ({ page, context }) => {
-		const response = await page.request.get(`${BASE_URL}/api/auth/me`);
-		if (response.ok()) {
-			test.skip(true, 'User is authenticated - this test verifies redirect for unauthenticated users');
-			return;
-		}
-
-		await context.clearCookies();
-
-		await page.goto(`${BASE_URL}/exercises`, { waitUntil: 'domcontentloaded', timeout: 30000 });
-		await page.waitForTimeout(2000);
-
-		await expect(page).toHaveURL(isAuthKitUrl);
-	});
-
-	test('complete login flow and verify authenticated state', async ({ page }) => {
-
-		const response = await page.request.get(`${BASE_URL}/api/auth/me`);
-		if (response.ok()) {
-			console.log('User is already authenticated via API');
-			
-			await page.goto(BASE_URL, { waitUntil: 'domcontentloaded', timeout: 30000 });
-			
-			await page.waitForTimeout(5000);
-			
-			const userAvatar = page.locator('button.rounded-full').first();
-			let isVisible = await userAvatar.isVisible().catch(() => false);
-			
-			if (isVisible) {
-				await expect(userAvatar).toBeVisible({ timeout: 5000 });
-				return;
-			}
-			
-			console.log('User avatar not visible, waiting for page to fully load...');
-			await page.waitForSelector('.animate-spin', { state: 'hidden', timeout: 15000 }).catch(() => {});
-			await page.waitForTimeout(3000);
-			
-			isVisible = await userAvatar.isVisible({ timeout: 5000 }).catch(() => false);
-			if (isVisible) {
-				await expect(userAvatar).toBeVisible({ timeout: 5000 });
-				return;
-			}
-			
-			console.log('User avatar still not visible - checking for sign in button');
-			const signInButton = page.locator('button:has-text("Sign In")').first();
-			const isSignInVisible = await signInButton.isVisible({ timeout: 3000 }).catch(() => false);
-			
-			if (isSignInVisible) {
-				await signInButton.click();
-				await expect(page).toHaveURL(isAuthKitUrl, { timeout: 15000 });
-			} else {
-				console.log('Neither avatar nor sign in visible - page may be in loading state');
-			}
-			return;
-		}
-
-		await page.goto(BASE_URL, { waitUntil: 'domcontentloaded', timeout: 30000 });
-		
-		await page.waitForTimeout(3000);
-
-		const currentUrl = page.url();
-		if (isAuthKitUrl(new URL(currentUrl))) {
-			console.log('Already on auth page, proceeding with login...');
-			
-			const emailInput = page.locator('input[name="email"]');
-			await emailInput.waitFor({ state: 'visible', timeout: 10000 });
-			
-			await emailInput.click();
-			await emailInput.fill(TEST_USERNAME);
-			await page.waitForTimeout(1000);
-			
-			const emailValue = await emailInput.inputValue();
-			console.log('Email value after fill:', emailValue ? `"${emailValue.substring(0, 5)}..."` : 'EMPTY');
-			
-			if (!emailValue || emailValue.length === 0) {
-				console.log('Filling with type instead...');
-				await emailInput.click({ clickCount: 3 });
-				await page.keyboard.type(TEST_USERNAME);
-				await page.waitForTimeout(1000);
-			}
-			
-			const continueBtn = page.locator('button:has-text("Continue")').first();
-			await continueBtn.click();
-			await page.waitForTimeout(5000);
-			
-			const passwordInput = page.locator('input[name="password"]');
-			const isPasswordVisible = await passwordInput.isVisible({ timeout: 5000 }).catch(() => false);
-			
-			if (!isPasswordVisible) {
-				console.log('Password field not visible, checking page state...');
-				const pageContent = await page.content();
-				if (pageContent.includes('Please enter your email')) {
-					console.log('Email validation error - trying again with explicit focus');
-					await emailInput.click();
-					await page.keyboard.press('Control+a');
-					await page.keyboard.type(TEST_USERNAME);
-					await page.waitForTimeout(500);
-					await continueBtn.click();
-					await page.waitForTimeout(5000);
-				}
-			}
-
-			await expect(page.locator('input[name="password"]')).toBeVisible({ timeout: 10000 });
-			await expect(page.locator('button[name="intent"]:not([data-method])')).toBeVisible();
-
-			await page.locator('input[name="password"]').fill(TEST_PASSWORD);
-			await page.locator('button[name="intent"]:not([data-method])').click();
-
-			await page.waitForURL(`${BASE_URL}/`, { timeout: 30000 });
-
-			const authCheck = await page.request.get(`${BASE_URL}/api/auth/me`);
-			expect(authCheck.ok()).toBe(true);
-
-			const userAvatar = page.locator('button.rounded-full').first();
-			await expect(userAvatar).toBeVisible({ timeout: 10000 });
-			return;
-		}
-
-		const signInButton = page.locator('button:has-text("Sign In")').first();
-		let isSignInVisible = await signInButton.isVisible({ timeout: 5000 }).catch(() => false);
-		
-		if (!isSignInVisible) {
-			console.log('Sign In button not visible, waiting for page to settle...');
-			await page.waitForSelector('.animate-spin', { state: 'hidden', timeout: 15000 }).catch(() => {});
-			await page.waitForTimeout(3000);
-			
-			isSignInVisible = await signInButton.isVisible({ timeout: 5000 }).catch(() => false);
-			
-			if (!isSignInVisible) {
-				console.log('Sign In button still not visible, checking if already authenticated...');
-				const userAvatar = page.locator('button.rounded-full').first();
-				const isAvatarVisible = await userAvatar.isVisible({ timeout: 5000 }).catch(() => false);
-				if (isAvatarVisible) {
-					console.log('User is already authenticated');
-					return;
-				}
-				throw new Error('Neither Sign In button nor user avatar visible after waiting');
-			}
-		}
-		
-		await signInButton.click();
-
-		await expect(page).toHaveURL(isAuthKitUrl, { timeout: 15000 });
-
-		const emailInput = page.locator(AUTH_EMAIL_SELECTOR);
-		await expect(emailInput).toBeVisible({ timeout: 10000 });
-
-		await emailInput.fill(TEST_USERNAME);
-		await page.locator(AUTH_CONTINUE_SELECTOR).click();
-
-		await expect(page.locator(AUTH_PASSWORD_SELECTOR)).toBeVisible({ timeout: 10000 });
-		await expect(page.locator(AUTH_SUBMIT_SELECTOR)).toBeVisible();
-
-		await page.locator(AUTH_PASSWORD_SELECTOR).fill(TEST_PASSWORD);
-		await page.locator(AUTH_SUBMIT_SELECTOR).click();
-
-		await page.waitForURL(`${BASE_URL}/`, { timeout: 30000 });
-
-		const authCheck = await page.request.get(`${BASE_URL}/api/auth/me`);
-		expect(authCheck.ok()).toBe(true);
-
-		const userAvatar = page.locator('button.rounded-full').first();
-		await expect(userAvatar).toBeVisible({ timeout: 10000 });
-	});
-
-	test('access protected routes after login', async ({ page }) => {
-		const authResponse = await page.request.get(`${BASE_URL}/api/auth/me`);
-		if (!authResponse.ok()) {
-			test.skip(true, 'User is not authenticated - this test requires prior authentication');
-			return;
-		}
-
-		await page.goto(`${BASE_URL}/exercises`, { waitUntil: 'networkidle' });
-		await expect(page).not.toHaveURL(isAuthKitUrl);
-
-		await page.goto(`${BASE_URL}/templates`, { waitUntil: 'networkidle' });
-		await expect(page).not.toHaveURL(isAuthKitUrl);
-
-		await page.goto(`${BASE_URL}/workouts`, { waitUntil: 'networkidle' });
-		await expect(page).not.toHaveURL(isAuthKitUrl);
-
-		await page.goto(`${BASE_URL}/history`, { waitUntil: 'networkidle' });
-		await expect(page).not.toHaveURL(isAuthKitUrl);
-	});
-
-	test('sign out and verify logged out state', async ({ page, context }) => {
-		const authResponse = await page.request.get(`${BASE_URL}/api/auth/me`);
-		if (!authResponse.ok()) {
-			test.skip(true, 'User is not authenticated - cannot test sign out');
-			return;
-		}
-
-		await page.goto(`${BASE_URL}/api/auth/signout`, { waitUntil: 'domcontentloaded', timeout: 30000 });
-
-		await page.waitForTimeout(5000);
-
-		await context.clearCookies();
-		await context.clearPermissions();
-
-		const afterSignOut = await page.request.get(`${BASE_URL}/api/auth/me`);
-		expect(afterSignOut.status()).toBe(401);
-
-		await page.goto(BASE_URL, { waitUntil: 'networkidle' });
-		await page.waitForTimeout(2000);
-
-		const signInButton = page.locator('button:has-text("Sign In")').first();
-		await expect(signInButton).toBeVisible({ timeout: 10000 });
-	});
-
-	test('re-authentication after logout redirects to protected route', async ({ page, context }) => {
-		const authResponse = await page.request.get(`${BASE_URL}/api/auth/me`);
-		if (!authResponse.ok()) {
-			test.skip(true, 'User is not authenticated - this test requires prior authentication');
-			return;
-		}
-
-		await page.goto(`${BASE_URL}/api/auth/signout`, { waitUntil: 'domcontentloaded', timeout: 30000 });
-		await page.waitForTimeout(3000);
-
-		await context.clearCookies();
-
-		const newPage = await context.newPage();
-		await newPage.goto(`${BASE_URL}/workouts/new`, { waitUntil: 'networkidle' });
-
-		await expect(newPage).toHaveURL(isAuthKitUrl, { timeout: 15000 });
-		await expect(newPage.locator(AUTH_EMAIL_SELECTOR)).toBeVisible({ timeout: 10000 });
-	});
-});
